# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "MIRI API",
  "description": "데이터 기반 상권 분석·업종 추천 서비스 MIRI의 REST API. 로그인 없이 누구나 쓰고, 저장 리포트는 브라우저마다 따로 보관한다(X-Device-Key 헤더). 금액 단위는 원(KRW).",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/analyses/{analysis_id}/reports": {
   "post": {
    "tags": [
     "분석 (SB-03~08)"
    ],
    "summary": "액션 리포트 생성 (UC-05)",
    "operationId": "post_report_api_v1_analyses__analysis_id__reports_post",
    "parameters": [
     {
      "name": "analysis_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Analysis Id"
      }
     },
     {
      "name": "X-Device-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "이 브라우저의 저장 키(없으면 저장 여부를 알려 주지 않음)",
       "title": "X-Device-Key"
      },
      "description": "이 브라우저의 저장 키(없으면 저장 여부를 알려 주지 않음)"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReportIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/reports/{report_id}": {
   "get": {
    "tags": [
     "리포트 (SB-08·09)"
    ],
    "summary": "액션 리포트 조회",
    "operationId": "get_report_api_v1_reports__report_id__get",
    "parameters": [
     {
      "name": "report_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Report Id"
      }
     },
     {
      "name": "X-Device-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "이 브라우저의 저장 키(없으면 저장 여부를 알려 주지 않음)",
       "title": "X-Device-Key"
      },
      "description": "이 브라우저의 저장 키(없으면 저장 여부를 알려 주지 않음)"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReportOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/reports/{report_id}/items/{action_id}": {
   "patch": {
    "tags": [
     "리포트 (SB-08·09)"
    ],
    "summary": "체크리스트 항목 완료 표시",
    "operationId": "set_item_done_api_v1_reports__report_id__items__action_id__patch",
    "parameters": [
     {
      "name": "report_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 36,
       "title": "Report Id"
      }
     },
     {
      "name": "action_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "maximum": 9223372036854775807,
       "minimum": 1,
       "title": "Action Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ActionDoneIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActionDoneOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/me/reports": {
   "post": {
    "tags": [
     "리포트 (SB-08·09)"
    ],
    "summary": "리포트 저장 (UC-06, 이 브라우저의 목록에)",
    "operationId": "save_report_api_v1_me_reports_post",
    "parameters": [
     {
      "name": "X-Device-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)",
       "title": "X-Device-Key"
      },
      "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SaveIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "리포트 (SB-08·09)"
    ],
    "summary": "이 브라우저의 저장 리포트 (SB-09)",
    "operationId": "my_reports_api_v1_me_reports_get",
    "parameters": [
     {
      "name": "X-Device-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)",
       "title": "X-Device-Key"
      },
      "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SavedListOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/me/reports/clear": {
   "post": {
    "tags": [
     "리포트 (SB-08·09)"
    ],
    "summary": "이 브라우저의 저장 리포트 모두 삭제(여럿이 쓰는 컴퓨터 정리)",
    "operationId": "clear_reports_api_v1_me_reports_clear_post",
    "parameters": [
     {
      "name": "X-Device-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)",
       "title": "X-Device-Key"
      },
      "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/me/reports/{saved_report_id}": {
   "delete": {
    "tags": [
     "리포트 (SB-08·09)"
    ],
    "summary": "저장 리포트 삭제",
    "operationId": "delete_report_api_v1_me_reports__saved_report_id__delete",
    "parameters": [
     {
      "name": "saved_report_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "maximum": 9223372036854775807,
       "minimum": 1,
       "title": "Saved Report Id"
      }
     },
     {
      "name": "report_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "minLength": 1,
         "maxLength": 36
        },
        {
         "type": "null"
        }
       ],
       "description": "이 저장 항목의 리포트 id(주면 일치할 때만 삭제)",
       "title": "Report Id"
      },
      "description": "이 저장 항목의 리포트 id(주면 일치할 때만 삭제)"
     },
     {
      "name": "X-Device-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)",
       "title": "X-Device-Key"
      },
      "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)"
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "리포트 (SB-08·09)"
    ],
    "summary": "저장 리포트 메모 수정(없으면 404 — 다시 저장하지 않음)",
    "operationId": "update_memo_api_v1_me_reports__saved_report_id__patch",
    "parameters": [
     {
      "name": "saved_report_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "maximum": 9223372036854775807,
       "minimum": 1,
       "title": "Saved Report Id"
      }
     },
     {
      "name": "X-Device-Key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)",
       "title": "X-Device-Key"
      },
      "description": "이 브라우저의 저장 키(영문·숫자·-·_ 32~128자)"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MemoIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "ActionDoneIn": {
    "properties": {
     "done": {
      "type": "boolean",
      "title": "Done",
      "description": "체크리스트 항목 완료 여부"
     }
    },
    "type": "object",
    "required": [
     "done"
    ],
    "title": "ActionDoneIn"
   },
   "ActionDoneOut": {
    "properties": {
     "action_id": {
      "type": "integer",
      "title": "Action Id"
     },
     "done": {
      "type": "boolean",
      "title": "Done"
     }
    },
    "type": "object",
    "required": [
     "action_id",
     "done"
    ],
    "title": "ActionDoneOut"
   },
   "ActionOut": {
    "properties": {
     "action_id": {
      "type": "integer",
      "title": "Action Id"
     },
     "action_type": {
      "type": "string",
      "title": "Action Type"
     },
     "content": {
      "type": "string",
      "title": "Content"
     },
     "priority": {
      "type": "integer",
      "title": "Priority"
     },
     "done": {
      "type": "boolean",
      "title": "Done",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "action_id",
     "action_type",
     "content",
     "priority"
    ],
    "title": "ActionOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MemoIn": {
    "properties": {
     "memo": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Memo"
     },
     "report_id": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 36,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Report Id",
      "description": "이 저장 항목의 리포트 id(주면 일치하는지 확인)"
     }
    },
    "type": "object",
    "title": "MemoIn"
   },
...
   "PolicyOut": {
    "properties": {
     "policy_id": {
      "type": "integer",
      "title": "Policy Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "provider": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Provider"
     },
     "support_type": {
      "type": "string",
      "title": "Support Type"
     },
     "summary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Summary"
     },
     "apply_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Apply Url"
     },
     "target_user": {
      "type": "string",
      "title": "Target User"
     },
     "target_region": {
      "type": "string",
      "title": "Target Region"
     },
     "checked_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Checked At"
     }
    },
    "type": "object",
    "required": [
     "policy_id",
     "name",
     "support_type",
     "target_user",
     "target_region"
    ],
    "title": "PolicyOut"
   },
...
   "ReportIn": {
    "properties": {
     "industry_code": {
      "type": "string",
      "title": "Industry Code"
     },
     "break_even_id": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 9.223372036854776e+18,
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Break Even Id"
     }
    },
    "type": "object",
    "required": [
     "industry_code"
    ],
    "title": "ReportIn"
   },
   "ReportOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "analysis_id": {
      "type": "string",
      "title": "Analysis Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "area_code": {
      "type": "string",
      "title": "Area Code"
     },
     "area_name": {
      "type": "string",
      "title": "Area Name"
     },
     "place_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Name"
     },
     "radius_m": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Radius M"
     },
     "industry_code": {
      "type": "string",
      "title": "Industry Code"
     },
     "industry_name": {
      "type": "string",
      "title": "Industry Name"
     },
     "one_line_summary": {
      "type": "string",
      "title": "One Line Summary"
     },
     "checklist": {
      "items": {
       "$ref": "#/components/schemas/ActionOut"
      },
      "type": "array",
      "title": "Checklist"
     },
     "policies": {
      "items": {
       "$ref": "#/components/schemas/PolicyOut"
      },
      "type": "array",
      "title": "Policies"
     },
     "risk": {
      "additionalProperties": true,
      "type": "object",
      "title": "Risk"
     },
     "breakeven": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Breakeven"
     },
     "recommendation": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Recommendation"
     },
     "saved_report_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Saved Report Id"
     },
     "data_quarter_label": {
      "type": "string",
      "title": "Data Quarter Label"
     }
    },
    "type": "object",
    "required": [
     "id",
     "analysis_id",
     "created_at",
     "area_code",
     "area_name",
     "industry_code",
     "industry_name",
     "one_line_summary",
     "checklist",
     "policies",
     "risk",
     "data_quarter_label"
    ],
    "title": "ReportOut"
   },
...
   "SaveIn": {
    "properties": {
     "report_id": {
      "type": "string",
      "maxLength": 36,
      "minLength": 1,
      "title": "Report Id"
     },
     "memo": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Memo"
     }
    },
    "type": "object",
    "required": [
     "report_id"
    ],
    "title": "SaveIn"
   },
   "SavedItemOut": {
    "properties": {
     "saved_report_id": {
      "type": "integer",
      "title": "Saved Report Id"
     },
     "report_id": {
      "type": "string",
      "title": "Report Id"
     },
     "analysis_id": {
      "type": "string",
      "title": "Analysis Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "area_name": {
      "type": "string",
      "title": "Area Name"
     },
     "place_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Name"
     },
     "industry_code": {
      "type": "string",
      "title": "Industry Code"
     },
     "industry_name": {
      "type": "string",
      "title": "Industry Name"
     },
     "suitability": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Suitability"
     },
     "risk_score": {
      "type": "integer",
      "title": "Risk Score"
     },
     "risk_grade": {
      "type": "string",
      "title": "Risk Grade"
     },
     "memo": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Memo"
     },
     "required_monthly_sales": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Required Monthly Sales",
      "description": "리포트에 반영한 손익분기 필요 월매출"
     },
     "monthly_fixed_cost": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Monthly Fixed Cost",
      "description": "월 고정비(반영한 손익분기, 없으면 분석 조건 — 목표 월수입 포함)"
     },
     "achievability_ratio": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Achievability Ratio"
     },
     "owner_included": {
      "type": "boolean",
      "title": "Owner Included",
      "description": "필요 월매출에 목표 월수입이 들어 있는지",
      "default": false
     },
     "edited": {
      "type": "boolean",
      "title": "Edited",
      "description": "반영한 손익분기에 분석 조건과 다르게 바꾼 값이 있는지",
      "default": false
     },
     "business_goal": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Business Goal",
      "description": "이 분석의 창업 목표(같은 업종 카드끼리 적합도가 다른 이유)"
     },
     "item_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Item Type",
      "description": "분석 결과에서 이 업종의 자리: TOP · INTEREST · CAUTION"
     },
     "rank": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rank"
     },
     "eligible": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Eligible",
      "description": "추천 조건을 통과했는지(관심 업종이면 추천 제외일 수 있음)"
     },
     "caution": {
      "type": "boolean",
      "title": "Caution",
      "description": "주의 업종(같은 업종 중 입지 위험 상위 10%)인지",
      "default": false
     },
     "sales_outlier": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sales Outlier",
      "description": "평균 매출 ÷ 서울 같은 업종 중간값(3배 이상일 때만 — 달성배율은 참고용)"
     },
     "checklist_total": {
      "type": "integer",
      "title": "Checklist Total",
      "default": 0
     },
     "checklist_done": {
      "type": "integer",
      "title": "Checklist Done",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "saved_report_id",
     "report_id",
     "analysis_id",
     "created_at",
     "area_name",
     "industry_code",
     "industry_name",
     "risk_score",
     "risk_grade"
    ],
    "title": "SavedItemOut"
   },
   "SavedListOut": {
    "properties": {
     "count": {
      "type": "integer",
      "title": "Count"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/SavedItemOut"
      },
      "type": "array",
      "title": "Items"
     }
    },
    "type": "object",
    "required": [
     "count",
     "items"
    ],
    "title": "SavedListOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}